// core/apiClient/api.ts
export interface Client {
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    organizationId: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    address?: string;
    /**
     * 
     * @type {LatLng}
     * @memberof Client
     */
    location?: LatLng;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    homePhone: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    cellPhone: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    email: string;
    /**
     * Extra client properties can be stored as a Map<String,any>
     * @type {any}
     * @memberof Client
     */
    extra: any;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    notes: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    updatedBy: string;
    /**
     * 
     * @type {number}
     * @memberof Client
     */
    createdAt: number;
    /**
     * 
     * @type {number}
     * @memberof Client
     */
    updatedAt: number;
    /**
     * 
     * @type {ClientOptions}
     * @memberof Client
     */
    options: ClientOptions;
}
/**
 * 
 * @export
 * @interface ClientOptions
 */
export interface ClientOptions {
    /**
     * 
     * @type {boolean}
     * @memberof ClientOptions
     */
    enableNotifications: boolean;
    /**
     * 
     * @type {string}
     * @memberof ClientOptions
     */
    primaryNotificationMethod: ClientOptions.PrimaryNotificationMethodEnum;
}

/**
 * @export
 * @namespace ClientOptions
 */
export namespace ClientOptions {
    /**
     * @export
     * @enum {string}
     */
    export enum PrimaryNotificationMethodEnum {
        SMS = <any> 'SMS',
        EMAIL = <any> 'EMAIL'
    }
}
// ...
export interface LatLng {
    /**
     * 
     * @type {number}
     * @memberof LatLng
     */
    latitude: number;
    /**
     * 
     * @type {number}
     * @memberof LatLng
     */
    longitude: number;
}